split utils_test.go
