+++ internal/utils/utils_test.go
@@ -1,36 +1,10 @@
 package utils
 
 import (
-	"log/slog"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/wasify-io/wasify-go/internal/memory"
 )
-
-func TestGetLogLevel(t *testing.T) {
-
-	newLogger := NewLogger(LogDebug)
-	assert.NotNil(t, newLogger)
-
-	tests := []struct {
-		severity LogSeverity
-		expected slog.Level
-	}{
-		{LogDebug, slog.LevelDebug},
-		{LogInfo, slog.LevelInfo},
-		{LogWarning, slog.LevelWarn},
-		{LogError, slog.LevelError},
-		{LogSeverity(255), slog.LevelInfo}, // Unexpected severity
-	}
-
-	for _, test := range tests {
-		got := GetlogLevel(test.severity)
-		if got != test.expected {
-			t.Errorf("for severity %d, expected %d but got %d", test.severity, test.expected, got)
-		}
-	}
-}
 
 func TestCalculateHash(t *testing.T) {
 	data := []byte("test")
